@@ -1,13 +1,13 @@
 [package]
 name = "kgm-decoder"
-version = "0.1.0"
+version = "0.1.3-dev"
 authors = ["TanZhou <tanzhou08@qq.com>"]
-edition = "2018"
+edition = "2024"
 repository = "https://github.com/ghtz08/kuguo-kgm-decoder"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-xz2 = "0.1.6"
-lazy_static = "1.4.0"
-clap = "2.33.3"
+xz2 = "0.1.7"
+infer = "0.19.0"
+clap = { version = "4.5.40", features = ["derive"] }
