@@ -1,6 +1,6 @@
-use lazy_static::lazy_static;
 use std::io::Read;
 use std::ops::Range;
+use std::sync::LazyLock;
 use xz2::read::XzDecoder;
 
 use super::Decoder;
@@ -17,25 +17,23 @@ impl<'a> KuGou<'a> {
     const PUB_KEY_LEN: u64 = 1170494464;
     const PUB_KEY_LEN_MAGNIFICATION: u64 = 16;
     const MAGIC_HEADER: [u8; 28] = [
-        0x7c, 0xd5, 0x32, 0xeb, 0x86, 0x02, 0x7f, 0x4b, 0xa8, 0xaf, 0xa6, 0x8e, 0x0f, 0xff,
-        0x99, 0x14, 0x00, 0x04, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
+        0x7c, 0xd5, 0x32, 0xeb, 0x86, 0x02, 0x7f, 0x4b, 0xa8, 0xaf, 0xa6, 0x8e, 0x0f, 0xff, 0x99,
+        0x14, 0x00, 0x04, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
     ];
 
     fn get_pub_key(index: Range<u64>) -> &'static [u8] {
         // TODO: 对 key 进行惰性解码（需要解决随之带来的静态变量线程安全问题）
-        static KGM_KEY_XZ: &[u8] = include_bytes!("../assets/kugou_key.xz");
-        lazy_static! {
-            static ref KEYS: Vec<u8> = (|| {
-                let mut xz_decoder = XzDecoder::new(Bytes::new(KGM_KEY_XZ));
-                let mut key =
-                    vec![0; (KuGou::PUB_KEY_LEN / KuGou::PUB_KEY_LEN_MAGNIFICATION) as usize];
-                if let Ok(_) = xz_decoder.read_exact(&mut key) {
-                    key
-                } else {
+        static KGM_KEY_XZ: &[u8] = include_bytes!("../../assets/kugou_key.xz");
+        static KEYS: LazyLock<Vec<u8>> = LazyLock::new(|| {
+            let mut xz_decoder = XzDecoder::new(Bytes::new(KGM_KEY_XZ));
+            let mut key = vec![0; (KuGou::PUB_KEY_LEN / KuGou::PUB_KEY_LEN_MAGNIFICATION) as usize];
+            match xz_decoder.read_exact(&mut key) {
+                Ok(_) => key,
+                _ => {
                     panic!("Failed to decode the KuGou key")
                 }
-            })();
-        }
+            }
+        });
 
         &KEYS[(index.start / KuGou::PUB_KEY_LEN_MAGNIFICATION) as usize
             ..(index.end / KuGou::PUB_KEY_LEN_MAGNIFICATION + 1) as usize]
@@ -135,7 +133,7 @@ struct Bytes<'a> {
 }
 
 impl<'a> Bytes<'a> {
-    fn new(data: &[u8]) -> Bytes {
+    fn new(data: &'a [u8]) -> Self {
         Bytes { data, pos: 0 }
     }
 }
@@ -153,8 +151,8 @@ impl Read for Bytes<'_> {
 
 #[test]
 fn test_decode() {
-    let mut decoder = KuGou::new(std::fs::File::open("src/assets/test_kugou_kgm.dat").unwrap());
-    let mut right_file = std::fs::File::open("src/assets/test_kugou_kgm_right.dat").unwrap();
+    let mut decoder = KuGou::new(std::fs::File::open("assets/test_kugou_kgm.dat").unwrap());
+    let mut right_file = std::fs::File::open("assets/test_kugou_kgm_right.dat").unwrap();
 
     let mut audio = Vec::new();
     let mut right_dat = Vec::new();
@@ -163,6 +161,6 @@ fn test_decode() {
     right_file.read_to_end(&mut right_dat).unwrap();
 
     println!("{} {}", audio.len(), right_dat.len());
-    
+
     assert!(audio == right_dat);
 }
